roman comments
