+++ x/twap/strategy_test.go
@@ -8,10 +8,103 @@ import (
 	"github.com/osmosis-labs/osmosis/v13/app/apptesting/osmoassert"
 	"github.com/osmosis-labs/osmosis/v13/osmomath"
 	"github.com/osmosis-labs/osmosis/v13/x/twap"
+	"github.com/osmosis-labs/osmosis/v13/x/twap/types"
 )
 
-// TestComputeArithmeticStrategyTwap tests computeArithmeticTwap on various inputs.
-// Contrary to computeTwap that handles the cases with zero delta correctly,
+type computeTwapTestCase struct {
+	startRecord    types.TwapRecord
+	endRecord      types.TwapRecord
+	twapStrategies []twap.TwapStrategy
+	quoteAsset     string
+	expTwap        sdk.Dec
+	expErr         bool
+	expPanic       bool
+}
+
+// TestComputeArithmeticTwap tests computeTwap on various inputs.
+// The test vectors are structured by setting up different start and records,
+// based on time interval, and their accumulator values.
+// Then an expected TWAP is provided in each test case, to compare against computed.
+func (s *TestSuite) TestComputeTwap() {
+	arithStrategy := &twap.ArithmeticTwapStrategy{
+		TwapKeeper: *s.App.TwapKeeper,
+	}
+
+	geomStrategy := &twap.GeometricTwapStrategy{
+		TwapKeeper: *s.App.TwapKeeper,
+	}
+
+	tests := map[string]computeTwapTestCase{
+		"arithmetic only, basic: spot price = 1 for one second, 0 init accumulator": {
+			startRecord: newOneSidedRecord(baseTime, sdk.ZeroDec(), true),
+			endRecord:   newOneSidedRecord(tPlusOne, OneSec, true),
+			quoteAsset:  denom0,
+			twapStrategies: []twap.TwapStrategy{
+				arithStrategy,
+			},
+			expTwap: sdk.OneDec(),
+		},
+		// this test just shows what happens in case the records are reversed.
+		// It should return the correct result, even though this is incorrect internal API usage
+		"arithmetic only: invalid call: reversed records of above": {
+			startRecord: newOneSidedRecord(tPlusOne, OneSec, true),
+			endRecord:   newOneSidedRecord(baseTime, sdk.ZeroDec(), true),
+			quoteAsset:  denom0,
+			twapStrategies: []twap.TwapStrategy{
+				arithStrategy,
+			},
+			expTwap: sdk.OneDec(),
+		},
+		"same record: denom0, end spot price = 0": {
+			startRecord: newOneSidedRecord(baseTime, sdk.ZeroDec(), true),
+			endRecord:   newOneSidedRecord(baseTime, sdk.ZeroDec(), true),
+			quoteAsset:  denom0,
+			twapStrategies: []twap.TwapStrategy{
+				arithStrategy,
+				geomStrategy,
+			},
+			expTwap: sdk.ZeroDec(),
+		},
+		"same record: denom1, end spot price = 1": {
+			startRecord: newOneSidedRecord(baseTime, sdk.ZeroDec(), true),
+			endRecord:   newOneSidedRecord(baseTime, sdk.ZeroDec(), true),
+			quoteAsset:  denom1,
+			twapStrategies: []twap.TwapStrategy{
+				arithStrategy,
+				geomStrategy,
+			},
+			expTwap: sdk.OneDec(),
+		},
+		"arithmetic only: accumulator = 10*OneSec, t=5s. 0 base accum": testCaseFromDeltas(
+			s, sdk.ZeroDec(), tenSecAccum, 5*time.Second, sdk.NewDec(2)),
+		"arithmetic only: accumulator = 10*OneSec, t=100s. 0 base accum (asset 1)": testCaseFromDeltasAsset1(s, sdk.ZeroDec(), OneSec.MulInt64(10), 100*time.Second, sdk.NewDecWithPrec(1, 1)),
+		"geometric only: accumulator = log(10)*OneSec, t=5s. 0 base accum": geometricTestCaseFromDeltas0(
+			s, sdk.ZeroDec(), geometricTenSecAccum, 5*time.Second, twap.TwapPow(geometricTenSecAccum.QuoInt64(5*1000))),
+		"geometric only: accumulator = log(10)*OneSec, t=100s. 0 base accum (asset 1)": geometricTestCaseFromDeltas1(s, sdk.ZeroDec(), geometricTenSecAccum, 100*time.Second, sdk.OneDec().Quo(twap.TwapPow(geometricTenSecAccum.QuoInt64(100*1000)))),
+		"three asset same record: asset1, end spot price = 1": {
+			startRecord: newThreeAssetOneSidedRecord(baseTime, sdk.ZeroDec(), true)[1],
+			endRecord:   newThreeAssetOneSidedRecord(baseTime, sdk.ZeroDec(), true)[1],
+			quoteAsset:  denom2,
+			expTwap:     sdk.OneDec(),
+			twapStrategies: []twap.TwapStrategy{
+				arithStrategy,
+				geomStrategy,
+			},
+		},
+	}
+	for name, test := range tests {
+		s.Run(name, func() {
+			for _, twapStrategy := range test.twapStrategies {
+				actualTwap, err := twap.ComputeTwap(test.startRecord, test.endRecord, test.quoteAsset, twapStrategy)
+				s.Require().NoError(err)
+				osmoassert.DecApproxEq(s.T(), test.expTwap, actualTwap, osmomath.GetPowPrecision())
+			}
+		})
+	}
+}
+
+// TestComputeArithmeticStrategyTwap tests arithmetic strategy's computeTwap
+// Contrary to computeTwap function (logic.go) that handles the cases with zero delta correctly,
 // this function should panic in case of zero delta.
 func (s *TestSuite) TestComputeArithmeticStrategyTwap() {
 	pointOneAccum := OneSec.QuoInt64(10)
@@ -64,6 +157,9 @@ func (s *TestSuite) TestComputeArithmeticStrategyTwap() {
 	}
 }
 
+// TestComputeGeometricStrategyTwap tests geometric strategy's computeTwap
+// Contrary to computeTwap function (logic.go) that handles the cases with zero delta correctly,
+// this function should panic in case of zero delta.
 func (s *TestSuite) TestComputeGeometricStrategyTwap() {
 	tests := map[string]computeTwapTestCase{
 		// basic test for both denom with zero start accumulator
@@ -137,7 +233,7 @@ func (s *TestSuite) TestComputeGeometricStrategyTwap() {
 	}
 }
 
-func (s *TestSuite) TestComputeArithmeticTwap_ThreeAsset_Arithmetic() {
+func (s *TestSuite) TestComputeArithmeticStrategyTwap_ThreeAsset() {
 	tenSecAccum := OneSec.MulInt64(10)
 	pointOneAccum := OneSec.QuoInt64(10)
 	tests := map[string]computeThreeAssetArithmeticTwapTestCase{
@@ -167,7 +263,7 @@ func (s *TestSuite) TestComputeArithmeticTwap_ThreeAsset_Arithmetic() {
 	}
 }
 
-func (s *TestSuite) TestComputeArithmeticTwap_ThreeAsset_Geometric() {
+func (s *TestSuite) TestComputeGeometricStrategyTwap_ThreeAsset() {
 	var (
 		five        = sdk.NewDec(5)
 		fiveFor3Sec = OneSec.MulInt64(3).Mul(twap.TwapLog(five))
